+++ R/data.R
@@ -1131,12 +1131,14 @@
 #'
 #' A string vector containing final demand sectors used for calculating Total Final Consumption (TFC)
 #' 
-#' @format A string list with `r length(fd_sectors)`
+#' @format A string list with `r length(fd_sectors)`, comprised of the following vectors of strings:
 #' \describe{
 #' \item{eiou_flows}{The string vector identifying energy industry own use flows.}
 #' \item{industry_net_flows}{The string vector identifying non-eiou (net) Industry flows.}
 #' \item{transport_domestic_flows}{The string vector identifying domestic transport flows.}
 #' \item{other_flows}{The string vector identifying Other flows.}
+#' \item{non_energy_flows}{The string vector of Non-energy flows.}
+#' \item{memo_non_energy_flows}{The string vector `memo_non_energy_flows` with leading "Memo: " stripped away.}
 #' }
 #' 
 #' @examples